+++ src/app/(auth)/SignUp/page.js
@@ -3,7 +3,7 @@ import Link from 'next/link';
 
 function Page() {
     return (
-        <div class="absolute inset-0 -z-10 h-full w-full items-center px-5 py-2 [background:radial-gradient(125%_125%_at_50%_10%,#000_40%,#3e6_100%)]">
+        <div className="absolute inset-0 -z-10 h-full w-full items-center px-5 py-2 [background:radial-gradient(125%_125%_at_50%_10%,#000_40%,#3e6_100%)]">
             <div className="flex min-h-full flex-col justify-center px-6 py-12 lg:px-8">
                 <div className="sm:mx-auto sm:w-full sm:max-w-sm">
                     <h2 className="mt-10 text-center text-3xl font-bold leading-9 tracking-tight text-white">Create Account</h2>
